+++ activerecord/test/models/hmt_enrollment.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class HmtEnrollment < ActiveRecord::Base
+  belongs_to :hmt_student
+  belongs_to :hmt_course
+end