+++ activerecord/test/models/hmt_student.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class HmtStudent < ActiveRecord::Base
+  has_many :hmt_enrollments
+  has_many :hmt_courses, through: :hmt_enrollments
+end